requirements.txt
langchain
langchain_community
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
langchain_openai
langchain-google-genai
transformers
huggingface_hub
matplotlib
seaborn
scikit-learn
pandas
numpy
chromadb 
google-genai